Remove promisifier and rename inject / background scripts to suitable names
