+++ chrome/extension/handleLiveChanges.js
