+++ webpack/prod.config.base.js
@@ -16,7 +16,7 @@ module.exports = (platform) => {
     entry: {
       popup: [customPath, path.join(__dirname, '../chrome/extension/popup')],
       background: [customPath, path.join(__dirname, '../chrome/extension/background')],
-      inject: [customPath, path.join(__dirname, '../chrome/extension/inject')]
+      handleLiveChanges: [customPath, path.join(__dirname, '../chrome/extension/handleLiveChanges')]
     },
     output: {
       path: path.join(__dirname, config.outputDir),
